+++ src/main/java/umc/cockple/demo/domain/notification/event/NotificationEvent.java
@@ -1,4 +1,4 @@
-package umc.cockple.demo.domain.notification.listener.event;
+package umc.cockple.demo.domain.notification.event;
 
 public record NotificationEvent(
         Long memberId,
