+++ src/main/java/umc/cockple/demo/domain/notification/service/NotificationV2CommandService.java
@@ -13,9 +13,9 @@ import umc.cockple.demo.domain.member.repository.MemberRepository;
 import umc.cockple.demo.domain.notification.domain.Notification;
 import umc.cockple.demo.domain.notification.domain.NotificationDestination;
 import umc.cockple.demo.domain.notification.domain.NotificationLegacyCompatibility;
-import umc.cockple.demo.domain.notification.dto.NotificationCreateCommand;
+import umc.cockple.demo.domain.notification.command.NotificationCreateCommand;
 import umc.cockple.demo.domain.notification.enums.NotificationResourceType;
-import umc.cockple.demo.domain.notification.listener.event.NotificationPushRequestedEvent;
+import umc.cockple.demo.domain.notification.event.NotificationPushRequestedEvent;
 import umc.cockple.demo.domain.notification.exception.NotificationErrorCode;
 import umc.cockple.demo.domain.notification.exception.NotificationException;
 import umc.cockple.demo.domain.notification.repository.NotificationRepository;
