+++ src/test/java/umc/cockple/demo/domain/notification/listener/NotificationEventListenerTest.java
@@ -17,7 +17,7 @@ import org.springframework.transaction.support.TransactionTemplate;
 import umc.cockple.demo.domain.member.domain.Member;
 import umc.cockple.demo.domain.member.repository.MemberRepository;
 import umc.cockple.demo.domain.notification.fcm.FcmService;
-import umc.cockple.demo.domain.notification.listener.event.NotificationEvent;
+import umc.cockple.demo.domain.notification.event.NotificationEvent;
 import umc.cockple.demo.global.enums.Gender;
 import umc.cockple.demo.global.enums.Level;
 import umc.cockple.demo.support.IntegrationTestBase;
